requirements.txt
streamlit
pandas
groq
python-dotenv
matplotlib
langchain
